Metod_Pollarda_Rebuild_2

+++ factorization/factorization/Algorithms_1.h
@@ -2,4 +2,4 @@
 #include <vector>
 using namespace std;
 
-vector <__int64> Pollardans(__int64 n);
+vector <int> Pollardans(int n);
