+++ factorization/factorization/MethodPollardaP_1_Nikitin.cpp
@@ -3,12 +3,12 @@
 #include <cmath>
 using namespace std;
 
-__int64 powmod(__int64 a, __int64 k, __int64 n) // (a*k) % n
+__int64 powmod(int a, int k, int n) // (a*k) % n
 {
-    return (__int64(pow(a, k)) % n);
+    return (int(pow(a, k)) % n);
 }
 
-bool prostoe(__int64 n)// проверяем простое число или нет
+bool prostoe(int n)// проверяем простое число или нет
 {
     for (unsigned int i = 2; i <= sqrt(n); i++) 
         if (n % i == 0)
@@ -16,27 +16,27 @@ bool prostoe(__int64 n)// проверяем простое число или н
     return true;
 }
 
-__int64 NOD(__int64 a, __int64 b) { //НОД по алгоритму Евклида
+int NOD(int a, int b) { //НОД по алгоритму Евклида
     return(a == 0) ? b : NOD(b % a, a);
 }
 
-__int64 metodpollardap1(__int64 n) {//реализация метода Полларда p-1
-    vector <__int64> prostch;
-    __int64 B;
+int metodpollardap1(int n) {//реализация метода Полларда p-1
+    vector <int> prostch;
+    int B;
     B = 10;
     for (unsigned int i = 2; i <= B; i++) {
         if (prostoe(i)) {
             prostch.push_back(i);
         }
     }
-    __int64 l;
-    __int64 a = 2;
-    __int64 p;
+    int l;
+    int a = 2;
+    int p;
     p = NOD(a, n);
     if (p == 1) {
         for (unsigned int i = 0; i <= prostch.size() - 1; i++) {
             if (NOD(a - 1, n) == 1) {
-                l = __int64(floor(log(B) / log(prostch[i])));
+                l = int(floor(log(B) / log(prostch[i])));
                 a = powmod(a, powmod(prostch[i], l, n), n);
                 if (a > 1)
                     p = NOD(a - 1, n);
@@ -46,9 +46,9 @@ __int64 metodpollardap1(__int64 n) {//реализация метода Полл
     return p;
 }
 
-vector <__int64> Pollardans(__int64 n) {//создаем вектор простых множителей числа
-    vector <__int64> mn;
-    __int64 p = 2;
+vector <int> Pollardans(int n) {//создаем вектор простых множителей числа
+    vector <int> mn;
+    int p = 2;
     while (n > 1 && p > 1) {
         if (!prostoe(n)) {
             p = metodpollardap1(n);
